Massive MPC refactoring
