+++ track/__init__.py
@@ -2,7 +2,6 @@ from .config import *
 from .gamepad import *
 from .mounts import *
 from .control import *
-from .errorsources import *
 from .telem import *
 from .laser import *
 from .plate_solve import *
